moved dict conversion to estimation module
